refactor: more `PeerSocket` to `connection_state` module
